gic: don't config eoimode if GIC_NOT_SUPPORT_DEACTIVATE defined
ICCICR.EOIMODES bit Controls the behavior of accesses to GICC_EOIR and GICC_DIR registers.
In GICv2,
0: GICC_EOIR has both priority drop and deactivate interrupt functionality.
   Accesses to the GICC_DIR are UNPREDICTABLE.
1: GICC_EOIR has priority drop functionality only.
   GICC_DIR has deactivate interrupt functionality

so if EOIMODE=1, where:
write to the ICCEOI register drops the priority of the interrupt and leave it active,
other interrupt at the same priority level can now be taken, but the active interrupt
cannot be taken again.
write to the ICCDIR marks the interrupts as inactive, meaning it can now be taken again.

Change-Id: I0f115387d9fece82a5ccfad30a1c972e8789e0ca
Signed-off-by: huangshr <huangshr@allwinnertech.com>